Renamed plugins: Now only a single plugin - $.tmpl and $.fn.tmpl. Renamed tmplcmd to tmplTags
Instance plugin:
Returns template context.

Static plugin:
For 1 param, passing in an html string or an element or jQuery instance for
a script block template declaration object, corresponds to previous $.tmpl
For more params, corresponds to original $.render plugin.
Corresponds to $.render, if you pass both a template (string/elem/jQuery instance)
and data.

Modified this pointer in callbacks to be template context, not data.
Removed index and Array features, which will be replaced by alternative API semantics.

Used jQuery data to store template contexts, to ensure disposal when DOM elements
are removed.

In the case of $("#someTmpl").tmpl(data).appendTo(selector) pattern this
required setting $.data on fragment that is not parented in HTML DOM.
Some oddities of behavior arise as a result of the default cacheable=false,
when inserting into DOM in domManip.
Consider eliminating this scenario and requiring alternative pattern:
$(selector).append("#someTmpl", data); (See next commit).