Moved some private helper functions out of inner function scopes, for memory optimization.
Added support for white space after target in tmpl tag, e.g. ${ foo } or
${ foo(x) }

Added support for targets that include '.'s and calls to functions, e.g.
${foo.toLowerCase()}

Added support for implicit passing of $data to nested tmpl when no
parameters, so that {{tmpl "#myTmpl"}} is equivalent to
{{tmpl() "#myTmpl"}}, to {{tmpl(null) "#myTmpl"}}, and to
{{tmpl($data) "#myTmpl"}}, and so allows myTmpl to bind to the same data
as the parent template.
(But {{tmpl(foo) "#myTmpl"}} or {{tmpl({}) "#myTmpl"}} will replace the
inherited $data value.)

Added support for inserting a template without data, using
$( "#tmpl" ).tmpl(null).appendTo( "#container" ); - which is equivalent to
$( "#tmpl" ).tmpl( {} ).appendTo( "#container" );