Fix to ignore null or undefined members of array data, when using .tmpl(dataArray) to render template against the array
Fix so that ${foo} when foo is a function will call foo with the
template item as 'this' pointer, rather than having the current data item
as 'this' and passing the template item as a parameter. This removes
dependencies on the signature of the function, so existing functions (not
designed to be called on some this object) will work.