Fixed issue 5 for parenting of {{wrap}} items, and associated issue of double events and potential memory leak.
Also a couple of other minor bug fixes.

Changed samples to always use parens when calling functions as template
tag parameters, since this is better practice. (Implicit function
invocation support only works in certain cases - basically when the
function is the only token in the expression, and not part of a more
complex expression. For this reason support for implicit function calls
may be removed later.)